+++ scripts/voucherHubUtils.ts
@@ -94,6 +94,6 @@ export async function getExpectedVoucherProxyCodeHash(voucherBeaconAddress: stri
          * Note: Look very carefully before updating this value to avoid messing with
          * existing vouchers already deployed in production.
          */
-        return '0x386d2735c0e7b2144a00e43bc5451acd8063573f30f1edc7b8e8891039cbd84f';
+        return '0xa1133afaac32e4a55b8ae263f5e5ca24950b033bc1dac2d1691ea4ebb3ddbf3c';
     }
 }